fix(web): resolve lint and e2e CI failures from Dusk rebrand
- Extract ThemeContext and useTheme into theme/themeContext.ts so
  ThemeProvider.tsx only exports a component, satisfying the
  react-refresh/only-export-components lint rule (was failing with
  --max-warnings=0).
- Scope the Learn search-input locator to .last() in the e2e spec.
  The new header GlobalSearch shares the "Search chapters..."
  placeholder with the sidebar search, which caused a Playwright
  strict-mode violation (2 matching elements). Mirrors the existing
  .last() pattern used for the "Open search" button.
