+++ apps/web/src/theme/ThemeProvider.tsx
@@ -1,22 +1,8 @@
-import React, {
-  createContext,
-  useCallback,
-  useContext,
-  useEffect,
-  useState,
-} from 'react';
+import React, { useCallback, useEffect, useState } from 'react';
 
-type Theme = 'light' | 'dark';
-
-interface ThemeContextValue {
-  theme: Theme;
-  setTheme: (theme: Theme) => void;
-  toggle: () => void;
-}
+import { ThemeContext, type Theme } from './themeContext';
 
 const STORAGE_KEY = 'athena-theme';
-
-const ThemeContext = createContext<ThemeContextValue | undefined>(undefined);
 
 /**
  * Resolve the initial theme from the source of truth: the `data-theme`
@@ -99,12 +85,4 @@ export const ThemeProvider: React.FC<{ children: React.ReactNode }> = ({
       {children}
     </ThemeContext.Provider>
   );
-};
-
-export const useTheme = (): ThemeContextValue => {
-  const ctx = useContext(ThemeContext);
-  if (!ctx) {
-    throw new Error('useTheme must be used within a ThemeProvider');
-  }
-  return ctx;
 };
