+++ apps/web/src/theme/themeContext.ts
@@ -1,0 +1,21 @@
+import { createContext, useContext } from 'react';
+
+export type Theme = 'light' | 'dark';
+
+export interface ThemeContextValue {
+  theme: Theme;
+  setTheme: (theme: Theme) => void;
+  toggle: () => void;
+}
+
+export const ThemeContext = createContext<ThemeContextValue | undefined>(
+  undefined,
+);
+
+export const useTheme = (): ThemeContextValue => {
+  const ctx = useContext(ThemeContext);
+  if (!ctx) {
+    throw new Error('useTheme must be used within a ThemeProvider');
+  }
+  return ctx;
+};
